requirements.txt
```bash
python-telegram-bot
psycopg2-binary
python-dotenv
